# Cpp version of torch tensor