--- a/R/RcppExports.R
+++ b/R/RcppExports.R
@@ -11265,6 +11265,10 @@ torch_tensor_cpp <- function(x, dtype, device, requires_grad, pin_memory) {
     .Call('_torch_torch_tensor_cpp', PACKAGE = 'torchpkg', x, dtype, device, requires_grad, pin_memory)
 }
 
+list_of_tensors <- function(x) {
+    .Call('_torch_list_of_tensors', PACKAGE = 'torchpkg', x)
+}
+
 cpp_as_array <- function(x) {
     .Call('_torch_cpp_as_array', PACKAGE = 'torchpkg', x)
 }