--- a/R/install.R
+++ b/R/install.R
@@ -1,4 +1,4 @@
-branch <- "libtorch1.10.2"
+branch <- "main"
 
 install_config <- list(
   "1.10.2" = list(