--- a/R/utils-data-collate.R
+++ b/R/utils-data-collate.R
@@ -1,13 +1,20 @@
+is_tensor_like <- function(x) {
+  # we want arrays, matrix's and vectors (lenght > 1) to be converted to
+  # tensors. this is only different from torch because there are native
+  # scalar types in Python
+  has_tensor_like_types <- is.atomic(x) && (!is.character(x))
+  has_tensor_like_types && (length(x) > 1 || is.array(x) || is.matrix(x))
+}
+
 utils_data_default_collate <- function(batch) {
   elem <- batch[[1]]
   if (is_torch_tensor(elem)) {
     return(torch_stack(batch, dim = 1))
-  } else if ((!is.character(elem) && !is.list(elem)) && is.atomic(elem) &&
-    (is.array(elem) || is.matrix(elem) || length(elem) > 1)) {
+  } else if (is_tensor_like(elem)) {
     # we want arrays, matrix's and vectors (lenght > 1) to be converted to
     # tensors. this is only different fromr torch because there's differentiation
     # between lenght 1 vectors and scalars.
-    return(utils_data_default_collate(list_of_tensors(batch)))
+    return(torch_stack(list_of_tensors(batch), dim = 1))
   } else if (is.integer(elem) && length(elem) == 1) {
     k <- unlist(batch)
     return(torch_tensor(k, dtype = torch_long()))