--- a/src/torch_api.cpp
+++ b/src/torch_api.cpp
@@ -128,10 +128,7 @@ XPtrTorchTensor from_sexp_tensor(SEXP x) {
 
   // TODO: it would be nice to make it all C++
   if (Rf_isVectorAtomic(x)) {
-    Rcpp::Environment torch_pkg = Rcpp::Environment("package:torch");
-    Rcpp::Function f = torch_pkg["torch_tensor"];
-    return XPtrTorchTensor(
-        Rcpp::as<Rcpp::XPtr<XPtrTorchTensor>>(f(x))->get_shared());
+    return torch_tensor_cpp(x);
   }
 
   Rcpp::stop("Expected a torch_tensor.");