Fix #470 

We also added a function `list_of_tensors` that casts a list of tensor_like objects to tensors, a little faster than `lapply(batch, torch_tensor)`.

With this PR we slightly improve performance of dataloaders. For instance:

New code:

``` r
data <- torchvision::cifar100_dataset(root = "./bench/", download = TRUE)
dl <- torch::dataloader(data, batch_size = 32)

bench::system_time({
  coro::loop(for (b in dl) {
    x <- b
  })
})
#> process    real 
#>   8.03s   8.08s
```

<sup>Created on 2022-02-11 by the [reprex package](https://reprex.tidyverse.org) (v2.0.1)</sup>

Old code:

``` r
data <- torchvision::cifar100_dataset(root = "./bench/", download = TRUE)
dl <- torch::dataloader(data, batch_size = 32)

bench::system_time({
  coro::loop(for (b in dl) {
    x <- b
  })
})
#> process    real 
#>   11.5s   11.5s
```

<sup>Created on 2022-02-11 by the [reprex package](https://reprex.tidyverse.org) (v2.0.1)</sup>